+++ zsh-alfred.sh
@@ -998,6 +998,11 @@ dualmonitor_stuffs() {
         'https://images4.alphacoders.com/204/204586.png'  # 2
         'https://www.dualmonitorbackgrounds.com/albums/SDuaneS/the-force-awakens-8.jpg'  # 3
         'https://www.dualmonitorbackgrounds.com/albums/SDuaneS/the-force-awakens-20.jpg'  # 4
+        'https://www.dualmonitorbackgrounds.com/albums/tonyjcooke/ifosbrasilflag3840x1080.jpg'  # 5
+        'https://www.dualmonitorbackgrounds.com/albums/tonyjcooke/tortolasunset2_3840x1080.jpg'  # 6
+        'https://www.dualmonitorbackgrounds.com/albums/SDuaneS/the-force-awakens-19.jpg'  # 7
+        'https://images2.alphacoders.com/502/502758.jpg'  # 8
+        'https://images5.alphacoders.com/574/574064.jpg'  # 9
     )
 
     local -a m=(
@@ -2432,27 +2437,35 @@ postgres_stuffs() {
 
         echo && read $'?\033[1;37mREBOOT IS REQUIRED. SHOULD I REBOOT NOW SIR? \n[Y/N] R: \033[m' option
 
-        # Or pg_createcluster 9.3 (version_number) main --start
+        # Or pg_createcluster $(apt version postgresql | cut -c1-2) main --start
         # /etc/init.d/postgresql start
-        for (( ; ; )); do
+        d+=(
+            /etc/postgresql/"$(apt version postgresql | cut -c1-2)"/main/conf.d/  # 3
+        )
 
-            if [[ "${option:0:1}" =~ ^(s|S|y|Y)$ ]] ; then
+        if [[ ! -d "${d[3]}" ]]; then
 
-                sudo reboot
+            for (( ; ; )); do
 
-            elif [[ "${option:0:1}" =~ ^(N|n)$ ]] ; then
+                if [[ "${option:0:1}" =~ ^(s|S|y|Y)$ ]] ; then
 
-                return_menu && break
+                    sudo reboot
 
-            else
+                elif [[ "${option:0:1}" =~ ^(N|n)$ ]] ; then
 
-                echo -ne ${c[RED]}"\n${e[flame]} SOME MEN JUST WANT TO WATCH THE WORLD BURN ${e[flame]}\n\t\t${c[WHITE]}PLEASE, ONLY Y OR N!\n\nSR. SHOULD I RESTART?${c[END]}\n${c[WHITE]}[Y/N] R: "${c[END]}
+                    return_menu && break
 
-                read option
+                else
 
-            fi
+                    echo -ne ${c[RED]}"\n${e[flame]} SOME MEN JUST WANT TO WATCH THE WORLD BURN ${e[flame]}\n\t\t${c[WHITE]}PLEASE, ONLY Y OR N!\n\nSR. SHOULD I RESTART?${c[END]}\n${c[WHITE]}[Y/N] R: "${c[END]}
 
-        done
+                    read option
+
+                fi
+
+            done
+
+        fi
 
     fi
